+++ crates/skytem_logs/src/mbed_motor_control/pid/header.rs
@@ -8,22 +8,22 @@ use byteorder::{LittleEndian, ReadBytesExt};
 use serde::{Deserialize, Serialize};
 use std::{fmt, io};
 use v1::PidLogHeaderV1;
-use v2::PidLogHeaderV2;
+use v2_beta::PidLogHeaderV2Beta;
 
 mod v1;
-mod v2;
+mod v2_beta;
 
 #[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
 pub(crate) enum PidLogHeader {
     V1(PidLogHeaderV1),
-    V2(PidLogHeaderV2),
+    V2Beta(PidLogHeaderV2Beta),
 }
 
 impl fmt::Display for PidLogHeader {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Self::V1(h) => write!(f, "{h}"),
-            Self::V2(h) => write!(f, "{h}"),
+            Self::V2Beta(h) => write!(f, "{h}"),
         }
     }
 }
@@ -61,13 +61,13 @@ impl PidLogHeader {
                 Self::V1(header)
             }
             2 => {
-                let (header, bytes_read) = PidLogHeaderV2::from_reader_with_uniq_descr_version(
+                let (header, bytes_read) = PidLogHeaderV2Beta::from_reader_with_uniq_descr_version(
                     reader,
                     unique_description,
                     version,
                 )?;
                 total_bytes_read += bytes_read;
-                Self::V2(header)
+                Self::V2Beta(header)
             }
             _ => {
                 return Err(io::Error::new(
