+++ test_data/mbed_motor_control/v2_beta/20241014_080729/status_20241014_080729_00.bin
